requirements.txt
lxml
python-dotenv
sqlalchemy
pymysql